- Scroll to element when its added
