+++ src/app/pages/submission/submission-edit/subm-sidebar/subm-check-sidebar/subm-check-sidebar.component.ts
@@ -2,6 +2,7 @@ import { FormControl, ValidationErrors } from '@angular/forms';
 import { Component, Input } from '@angular/core';
 import { ServerError } from 'app/shared/server-error.handler';
 import { SubmEditService } from '../../shared/subm-edit.service';
+import { scrollToFormControl } from 'app/utils';
 
 @Component({
   selector: 'st-subm-check-sidebar',
@@ -56,20 +57,7 @@ export class SubmCheckSidebarComponent {
    * @param {FieldControl} control - Form control augmented with the DOM element for the field.
    */
   onReviewClick(control: FormControl) {
-    const element: HTMLElement  = (<any>control).nativeElement;
-
-    if (element !== undefined) {
-      const rect = element.getBoundingClientRect();
-
-      if (!this.isInViewPort(rect)) {
-        window.scrollBy(0, rect.top - 120);
-      }
-
-      const elementRef: HTMLTextAreaElement | HTMLInputElement =
-        element.querySelectorAll<HTMLTextAreaElement | HTMLInputElement>('input, textarea')[0];
-
-      elementRef.focus();
-    }
+    scrollToFormControl(control);
   }
 
   /**
@@ -90,14 +78,5 @@ export class SubmCheckSidebarComponent {
       return 'not unique';
     }
     return Object.keys(errors)[0];
-  }
-
-  private isInViewPort(rect: { bottom: number, left: number, right: number, top: number }) {
-    return (
-      rect.top >= 0 &&
-      rect.left >= 0 &&
-      rect.bottom <= (window.innerHeight || document.documentElement!.clientHeight) &&
-      rect.right <= (window.innerWidth || document.documentElement!.clientWidth)
-    );
   }
 }
