+++ src/app/pages/submission/submission-edit/subm-sidebar/subm-edit-sidebar/subm-edit-sidebar.component.html
@@ -52,7 +52,6 @@
         Cancel
       </button>
     </li>
-
   </form>
 
   <li *ngIf="isEditModeOff && userData.isPrivileged && isAdvancedVisible" class="text-right">
