+++ src/app/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './date.utils';
 export * from './orcid.utils';
+export * from './scroll.utils';
